Fix removal of builds and branches
When the BitbucketServer returns something that isn't valid json for the
get methods in BitbucketServerAPIClient those methods were returning
null. This is effectively interpreted in calling methods as there being
none of theobjects we were looking for thus causing branches, PRs etc
to be trimmed erroneously. We need to throw a exception instead.